+++ R/every-some-none.R
@@ -1,10 +1,11 @@
-#' Do every or some elements of a list satisfy a predicate?
+#' Do every, some, or none of the elements of a list satisfy a predicate?
 #'
 #' @inheritParams map
 #' @param .p A predicate function to apply on each element of `.x`.
 #'   `some()` returns `TRUE` when `.p` is `TRUE` for at least one
 #'   element. `every()` returns `TRUE` when `.p` is `TRUE` for all
-#'   elements.
+#'   elements. `none()` returns `TRUE` when `.p` is `FALSE` for all
+#'   elements.`
 #' @param ... Additional arguments passed on to `.p`.
 #' @return A logical vector of length 1.
 #' @export
@@ -12,6 +13,8 @@
 #' y <- list(0:10, 5.5)
 #' y %>% every(is.numeric)
 #' y %>% every(is.integer)
+#' y %>% some(is.integer)
+#' y %>% none(is.character)
 every <- function(.x, .p, ...) {
   .p <- as_predicate(.p, ..., .mapper = TRUE, .deprecate = TRUE)
 
@@ -36,4 +39,10 @@ some <- function(.x, .p, ...) {
   }
 
   val
+}
+
+#' @export
+#' @rdname every
+none <- function(.x, .p, ...) {
+  every(.x, negate(.p), ...)
 }